Fixes tolower usage with cast to int
git-svn-id: svn+ssh://svn.code.sf.net/p/cmusphinx/code/trunk/SphinxTrain@9060 94700074-3cef-4d97-a70e-9c8c206c02f5

+++ src/libs/libcommon/lts.c
@@ -81,7 +81,7 @@ int lts_apply(const char *in_word,const char *feats,
     word = ckd_salloc((char *)in_word);
     if (!r->letter_table)
 	for (i = 0; i < strlen(word); ++i)
-	    word[i] = tolower(word[i]);
+	    word[i] = tolower((int)word[i]);
 
     /* Fill in out_phones structure as best we can. */
     maxphones = strlen(word) + 10;
